refactor(router): move screen options from layouts into screen files
