+++ src/app/(app)/style.tsx
@@ -1,3 +1,4 @@
+import { Tabs } from 'expo-router';
 import * as React from 'react';
 
 import { Buttons } from '@/components/buttons';
@@ -5,10 +6,18 @@ import { Colors } from '@/components/colors';
 import { Inputs } from '@/components/inputs';
 import { Typography } from '@/components/typography';
 import { FocusAwareStatusBar, SafeAreaView, ScrollView } from '@/components/ui';
+import { Style as StyleIcon } from '@/components/ui/icons';
 
 export default function Style() {
   return (
     <>
+      <Tabs.Screen
+        options={{
+          title: 'Style',
+          tabBarIcon: ({ color }) => <StyleIcon color={color} />,
+          tabBarButtonTestID: 'style-tab',
+        }}
+      />
       <FocusAwareStatusBar />
       <ScrollView className="px-4">
         <SafeAreaView className="flex-1">
