+++ src/app/www.tsx
@@ -1,6 +1,5 @@
-import { useLocalSearchParams, useNavigation, useRouter } from 'expo-router';
+import { Stack, useLocalSearchParams, useRouter } from 'expo-router';
 import * as React from 'react';
-import { useEffect } from 'react';
 import { View } from 'react-native';
 import { WebView } from 'react-native-webview';
 
@@ -10,19 +9,16 @@ import { translate } from '@/lib';
 export default function WWW() {
   const router = useRouter();
   const { url, title } = useLocalSearchParams();
-  const navigation = useNavigation();
 
-  useEffect(() => {
-    if (title !== undefined) {
-      navigation.setOptions({
-        title,
-      });
-    }
-  }, [navigation, title]);
+  const screenOptions = {
+    presentation: 'modal',
+    title: typeof title === 'string' ? title : '',
+  } as const;
 
   if (url === undefined || typeof url !== 'string') {
     return (
       <View className="flex-1 items-center justify-center bg-white">
+        <Stack.Screen options={screenOptions} />
         <Text className="text-lg text-red-500">
           {translate('www.invalidUrl')}
         </Text>
@@ -32,6 +28,7 @@ export default function WWW() {
 
   return (
     <View className="flex-1 bg-white">
+      <Stack.Screen options={screenOptions} />
       <WebView
         source={{ uri: url }}
         className="flex-1"
